Edit AboutUs.md
The AboutUs is the sample provided and it can be confusing when the user
like to contact the developers for Networkers.

Add my profile into the AboutUs file.

Edit UserGuide.md

The User Guide is the sample provided and it can be confusing to follow
 when it does not reflect the current product.

Change and edit the User Guide to reflect the one written in the google
docs.

Add teominghuang.md

We are trying to get used to creating a skeletal PPP. This allows
employees and the module to track user contribution.

Add my own profile and contribution into the skeletal PPP to get used
to the feature.