Create Completion and ClientInfo attribute
There was no way to indicate completion status prior.
With this, there is now an attribute to signal
Procedure completion.

There was also no way to indicate the client
whom the procedure belongs to earlier. With this
commit, there will be a backwards link to the client
from the Procedure.