Update LocalDate to LocalDateTime
In the previous PR, the comparison between the two Date classes for the autosort feature was between LocalDate.

However, in a later update to master, we have updated the Date class to use LocalDateTime instead. Hence, this PR is to update the autosort feature to reflect that.

This PR also updated the compareTo to compare the Date class with another Date object as commented by @iztanpy.