Create EditProcCommand
This is part of the component that will be
used to allow edit of command.

This is important in providing intuition to the
user when they have created a procedure and
will like that editted.

Create a EditProcCommand that provides foundation
to the feature.

Complete this first as it provides the crucial
component to the larger feature.

The feature is incomplete and this feature will be
completed in the following few commits.